Add business logic layer.
