+++ app.py
@@ -1,25 +1,10 @@
 from flask import Flask
-from api.endpoints.accounts import Accounts
 
+# Create the Flask application
 app = Flask(__name__)
 
-accounts = Accounts()
-
-
-@app.route('/reset', methods=['POST'])
-def reset():
-    return accounts.reset_state()
-
-
-@app.route('/balance', methods=['GET'])
-def get_balance():
-    return accounts.get_balance()
-
-
-@app.route('/event', methods=['POST'])
-def handle_events():
-    return accounts.process_event()
-
+# Import the endpoints
+from api.endpoints.accounts import *
 
 if __name__ == "__main__":
     app.run(debug=True)
